gtk: Manually implement GtkAlign
Since GTK 4.11.x a GTK_BASELINE_FILL was added pointing to GTK_BASELINE
and somehow gir doesn't know what to do here and just omits the
GTK_BASELINE from being generated.

To avoid that, manually implement that. Ideally, one would fix gir but
it is too much for a niche case that i won't bother doing so now.
